+++ cmd/cmd_internal_test.go
@@ -1,4 +1,4 @@
-package main
+package cmd
 
 import (
 	"os"
@@ -25,7 +25,8 @@ func TestRun(t *testing.T) {
 			args = append(args, p)
 
 			require.NotPanics(t, func() {
-				run(args)
+				// TODO: handle errors
+				_ = Run(args)
 			})
 		}
 	}
@@ -39,7 +40,7 @@ func TestRun(t *testing.T) {
 		for _, p := range pats {
 			os.Args = p
 
-			require.NotNilf(t, run(os.Args), "#%v", p)
+			require.NotNilf(t, Run(os.Args), "#%v", p)
 		}
 	}
 
@@ -53,7 +54,7 @@ func TestRun(t *testing.T) {
 		for _, p := range pats {
 			os.Args = p
 
-			require.NotNilf(t, run(os.Args), "#%v", p)
+			require.NotNilf(t, Run(os.Args), "#%v", p)
 		}
 	}
 
@@ -67,7 +68,20 @@ func TestRun(t *testing.T) {
 		for _, p := range pats {
 			os.Args = p
 
-			require.NotNilf(t, run(os.Args), "#%v", p)
+			require.NotNilf(t, Run(os.Args), "#%v", p)
+		}
+	}
+
+	{
+		// Fail case for `version` task.
+		pats := [][]string{
+			{"version", "--dir", "non-existent"},
+		}
+
+		for _, p := range pats {
+			os.Args = p
+
+			require.NotNilf(t, Run(os.Args), "#%v", p)
 		}
 	}
 }